refactor: replace business logic to services layer
