+++ src/artists/artists.service.ts
@@ -1,16 +1,30 @@
 import {
   BadRequestException,
+  Inject,
   Injectable,
   NotFoundException,
+  forwardRef,
 } from '@nestjs/common';
 
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { Artist } from './entities/artist.entity';
+import { AlbumsService } from '../albums/albums.service';
+import { TracksService } from '../tracks/tracks.service';
+import { FavoritesService } from '../favorites/favorites.service';
 
 @Injectable()
 export class ArtistsService {
   artists: Record<string, Artist> = {};
+
+  constructor(
+    @Inject(forwardRef(() => TracksService))
+    private tracksService: TracksService,
+    @Inject(forwardRef(() => AlbumsService))
+    private albumsService: AlbumsService,
+    @Inject(forwardRef(() => FavoritesService))
+    private favoritesService: FavoritesService,
+  ) {}
 
   findAll() {
     return Object.keys(this.artists).map((artistId) => this.artists[artistId]);
@@ -49,6 +63,21 @@ export class ArtistsService {
   }
 
   remove(artistId: string) {
+    const album = this.albumsService.findAlbumByArtistId(artistId);
+    const track = this.tracksService.findTrackByArtistId(artistId);
+
+    if (album) {
+      this.albumsService.update(album.id, { artistId: null });
+    }
+
+    if (track) {
+      this.tracksService.update(track.id, { artistId: null });
+    }
+
+    if (this.favoritesService.isArtistExists(artistId)) {
+      this.favoritesService.deleteArtist(artistId);
+    }
+
     const artist = this.findOne(artistId);
 
     delete this.artists[artist.id];
