+++ src/tracks/tracks.service.ts
@@ -1,16 +1,30 @@
 import {
   BadRequestException,
+  Inject,
   Injectable,
   NotFoundException,
+  forwardRef,
 } from '@nestjs/common';
 
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
 import { Track } from './entities/track.entity';
+import { AlbumsService } from '../albums/albums.service';
+import { ArtistsService } from '../artists/artists.service';
+import { FavoritesService } from '../favorites/favorites.service';
 
 @Injectable()
 export class TracksService {
   tracks: Record<string, Track> = {};
+
+  constructor(
+    @Inject(forwardRef(() => AlbumsService))
+    private albumsService: AlbumsService,
+    @Inject(forwardRef(() => ArtistsService))
+    private artistsService: ArtistsService,
+    @Inject(forwardRef(() => FavoritesService))
+    private favoritesService: FavoritesService,
+  ) {}
 
   findAll() {
     return Object.keys(this.tracks).map((trackId) => this.tracks[trackId]);
@@ -27,6 +41,13 @@ export class TracksService {
   }
 
   create({ name, albumId, artistId, duration }: CreateTrackDto) {
+    if (artistId) {
+      this.artistsService.findOne(artistId);
+    }
+    if (albumId) {
+      this.albumsService.findOne(albumId);
+    }
+
     const track = new Track(name, albumId, artistId, duration);
     this.tracks[track.id] = track;
 
@@ -34,6 +55,16 @@ export class TracksService {
   }
 
   update(trackId: string, updateTrackDto: UpdateTrackDto) {
+    const { artistId, albumId } = updateTrackDto;
+
+    if (artistId) {
+      this.artistsService.findOne(artistId);
+    }
+
+    if (albumId) {
+      this.albumsService.findOne(albumId);
+    }
+
     const hasProperties = Object.keys(updateTrackDto).length > 0;
 
     if (!hasProperties) {
@@ -49,6 +80,10 @@ export class TracksService {
   }
 
   remove(trackId: string) {
+    if (this.favoritesService.isTrackExists(trackId)) {
+      this.favoritesService.deleteTrack(trackId);
+    }
+
     const track = this.findOne(trackId);
 
     delete this.tracks[track.id];
